+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerGitRepoSearchFilters.java
@@ -8,7 +8,9 @@ public record ContainerGitRepoSearchFilters(
 
         List<String> includeExtensionFiles,
 
-        List<String> excludeExtensionFiles
+        List<String> excludeExtensionFiles,
+
+        String patternForIncludeFile
 
 ) {
 }