+++ fugitive-darkness-provider/src/main/java/com/zer0s2m/fugitivedarkness/provider/impl/SearchEngineGitGrepImpl.java
@@ -85,6 +85,10 @@ class SearchEngineGitGrepImpl extends SearchEngineGitGrepAbstract implements Sea
                             continue;
                         }
 
+                        if (!getWhetherSearchByIncludeFileByPattern(it.getEntryPathString())) {
+                            continue;
+                        }
+
                         if (getWhetherSearchByIncludeFileExtension(extensionFile)) {
                             List<ContainerInfoSearchFileMatcherGitRepo> matchers = getMatchedLines(
                                     objectLoader.openStream(), it.getEntryPathString());